test(core): split watch plugin order tests
